azure-functions/requirements 2.txt
azure-functions
-e ../